DRY-principle optimization
